+++ .gitignore
@@ -16,3 +16,6 @@ bakery
 baker
 function.zip
 main
+
+# local file for setting envvars '. ./set_vars.sh'
+set_vars.sh
