+++ Makefile
@@ -2,6 +2,7 @@
 GOCMD=go
 GOBUILD=$(GOCMD) build
 GOCLEAN=$(GOCMD) clean
+GORUN=$(GOCMD) run
 GOTEST=$(GOCMD) test
 GOGET=$(GOCMD) get
 BINARY_NAME=bakery
@@ -14,13 +15,11 @@ build_lambda:
 				GOOS=linux $(GOBUILD) -o main -v
 				zip function.zip main
 test: 
-				$(GOTEST) -v ./...
+				$(GOTEST) -v -race -count=1 ./...
 clean: 
 				$(GOCLEAN)
 				rm -f $(BINARY_NAME)
 				rm -f $(BINARY_UNIX)
 run:
-				$(GOGET)
-				$(GOBUILD) -o $(BINARY_NAME)
-				./$(BINARY_NAME)
+				$(GORUN) cmd/http/main.go
