+++ go.mod
@@ -1,0 +1,13 @@
+module github.com/cbsinteractive/bakery
+
+replace github.com/zencoder/go-dash => github.com/cbsinteractive/go-dash v0.0.0-20200107194942-8356b3b1db96
+
+go 1.13
+
+require (
+	github.com/akrylysov/algnhsa v0.12.1
+	github.com/grafov/m3u8 v0.11.1
+	github.com/kelseyhightower/envconfig v1.4.0
+	github.com/sirupsen/logrus v1.4.2
+	github.com/zencoder/go-dash v0.0.0-00010101000000-000000000000
+)
