+++ pkg/filters/filter.go
@@ -1,6 +1,8 @@
 package filters
 
-import "github.com/cbsinteractive/bakery/parsers"
+import (
+	"github.com/cbsinteractive/bakery/pkg/parsers"
+)
 
 // Filter is an interface for HLS and DASH filters
 type Filter interface {
